+++ src/corpus/CorpusOverview.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { PhPencilSimple } from "@phosphor-icons/vue";
+import useConfig from "./config/config.composable";
 import useCorpusIdParam from "@/corpus/corpusIdParam.composable";
-import { useCorpusState } from "@/corpus/corpusState.composable";
 import ConfigPanel from "@/corpus/config/ConfigPanel.vue";
 import SourcesPanel from "@/corpus/sources/SourcesPanel.vue";
 import JobStatus from "@/corpus/job/JobStatus.vue";
@@ -12,7 +12,7 @@ import RouteButton from "@/components/RouteButton.vue";
 import LayoutBox from "@/components/LayoutBox.vue";
 
 const corpusId = useCorpusIdParam();
-const { isNeedingConfig, isNeedingMeta } = useCorpusState(corpusId);
+const { hasMetadata, isConfigValid } = useConfig(corpusId);
 </script>
 
 <template>
@@ -27,7 +27,7 @@ const { isNeedingConfig, isNeedingMeta } = useCorpusState(corpusId);
         <template #controls>
           <RouteButton
             :to="`/library/corpus/${corpusId}/config`"
-            :class="{ 'button-primary': isNeedingConfig || isNeedingMeta }"
+            :class="{ 'button-primary': !isConfigValid || !hasMetadata }"
           >
             <PhPencilSimple weight="bold" class="inline mb-1 mr-1" />
             {{ $t("edit") }}