+++ src/corpus/analysis/AnalysisPanel.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { computed, ref, watch } from "vue";
 import { PhDownloadSimple, PhGearFine, PhInfo } from "@phosphor-icons/vue";
-import { useCorpusState } from "../corpusState.composable";
 import useExports from "../exports/exports.composable";
 import useJob from "../job/job.composable";
+import useConfig from "../config/config.composable";
 import ActionButton from "@/components/ActionButton.vue";
 import PendingContent from "@/spin/PendingContent.vue";
 
@@ -11,17 +11,20 @@ const props = defineProps<{
   corpusId: string;
 }>();
 
+const { isConfigValid, hasMetadata } = useConfig(props.corpusId);
 const { runJob, clearAnnotations, jobState, isJobRunning } = useJob(
   props.corpusId,
 );
-const { canBeReady } = useCorpusState(props.corpusId);
 const { exports, loadExports, downloadResult, getDownloadFilename } =
   useExports(props.corpusId);
-const { isDone } = useCorpusState(props.corpusId);
 
 const isPending = ref(false);
 const canRun = computed(
-  () => canBeReady.value && !isPending.value && !isJobRunning.value,
+  () =>
+    isConfigValid.value &&
+    hasMetadata.value &&
+    !isPending.value &&
+    !isJobRunning.value,
 );
 
 loadExports();
@@ -33,8 +36,8 @@ async function doRunJob() {
 }
 
 // When a job finishes, show download button.
-watch(isDone, () => {
-  if (isDone.value) {
+watch(jobState, () => {
+  if (jobState.value?.sparv == "done") {
     loadExports();
   }
 });