+++ src/corpus/job/JobStatus.vue
@@ -4,7 +4,6 @@ import useJob from "@/corpus/job/job.composable";
 import JobStatusMessage from "@/corpus/job/JobStatusMessage.vue";
 import { formatDate } from "@/util";
 import PendingContent from "@/spin/PendingContent.vue";
-import { useCorpusState } from "@/corpus/corpusState.composable";
 import ActionButton from "@/components/ActionButton.vue";
 import ProgressBar from "@/components/ProgressBar.vue";
 import TextData from "@/components/TextData.vue";
@@ -13,10 +12,9 @@ const props = defineProps<{
   corpusId: string;
 }>();
 
-const { abortJob, jobStatus, isJobRunning } = useJob(props.corpusId);
-const { isFailed } = useCorpusState(props.corpusId);
+const { abortJob, jobStatus, isJobRunning, hasError } = useJob(props.corpusId);
 
-const hasStarted = computed(
+const isStarted = computed(
   () =>
     Object.values(jobStatus.value?.status || {}).some(
       (status) => status != "none",
@@ -52,7 +50,7 @@ const hasStarted = computed(
       class="w-full my-2"
     />
 
-    <table v-if="hasStarted" class="w-full">
+    <table v-if="isStarted" class="w-full">
       <thead></thead>
       <tbody>
         <tr v-if="jobStatus.errors">
@@ -73,10 +71,10 @@ const hasStarted = computed(
           </td>
         </tr>
 
-        <tr v-if="isFailed && jobStatus.sparv_output">
+        <tr v-if="hasError && jobStatus.sparv_output">
           <th colspan="2">{{ $t("sparvOutput") }}</th>
         </tr>
-        <tr v-if="isFailed && jobStatus.sparv_output">
+        <tr v-if="hasError && jobStatus.sparv_output">
           <td colspan="2">
             <TextData :text="jobStatus.sparv_output" class="mb-2" />
           </td>